unix transport: Fix premature thread exit on initialization.
The 'initialized' value may still be 0 inside the thread, which would
cause it to exit prematurely.  Set 'initialized' to 1 before starting
the thread to fix this race condition.

This commit is similar to commit 817a7e6588 ("Fixed race condition when
starting WebSockets threads and initialized still 0") but for the pfunix
transport.